tapdb: modify type of transfer ID foreign key in burn table
This aligns the foreign key to be the same type as the referenced
table's primary key. This was an oversight that resulted from our type
replacement map we use to change the types of our primary keys to be
compatible with what SQLite and Postgres support.
So even though all our primary keys in the SQL files are defined as
INTEGER PRIMARY KEY, they are rewritten to BIGSERIAL PRIMARY KEY in
Postgres to allow auto increment, which corresponds to a BIGINT.
See scripts/gen_sqlc_docker.sh for a more in-depth explanation.

+++ tapdb/assets_store_test.go
@@ -2865,7 +2865,7 @@ func TestQueryAssetBurns(t *testing.T) {
 	assetID := inputAsset.ID()
 
 	_, err = assetsStore.db.InsertBurn(ctx, sqlc.InsertBurnParams{
-		TransferID: int32(assetTransfers[0].ID),
+		TransferID: assetTransfers[0].ID,
 		Note: sql.NullString{
 			String: "burn",
 			Valid:  true,
@@ -2883,7 +2883,7 @@ func TestQueryAssetBurns(t *testing.T) {
 	require.Len(t, burns, 1)
 
 	_, err = assetsStore.db.InsertBurn(ctx, sqlc.InsertBurnParams{
-		TransferID: int32(assetTransfers[0].ID),
+		TransferID: assetTransfers[0].ID,
 		Note: sql.NullString{
 			String: "burn",
 			Valid:  true,
